savepoint: move command handle logic to command files
